# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "EP1 Exam",
    "description": "documentation for Noroff EP - e-commerce"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "Logs user or admin into application, provides JWT token for 2 hours",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Registration"
        ],
        "description": "Registers new user to the application",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "description": "Gets list of all brands",
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "description": "Gets list of all brands",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/postbrand"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Brands"
        ],
        "description": "Gets brand of specified id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/alterbrand"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Brands"
        ],
        "description": "Gets brand of specified id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/delete"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/brands/{brandid}": {
      "get": {
        "tags": [
          "Brands"
        ],
        "description": "Gets brand of specified id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "brandid",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Gets list of all products",
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Creates new product",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/postproduct"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "description": "Alters existing product",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/alterproduct"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "description": "Soft deletes product of provided id(sets Processed : 1)",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/delete"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/{productid}": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Gets product of provided id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "productid",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "login": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "email or username"
        },
        "password": {
          "type": "string",
          "example": "YouS@feP@ssw0rd"
        }
      },
      "required": [
        "name",
        "password"
      ]
    },
    "register": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "Shoresy"
        },
        "lastName": {
          "type": "string",
          "example": "Waffles"
        },
        "email": {
          "type": "string",
          "example": "shoresy69@sudbluebulldogs.com"
        },
        "password": {
          "type": "string",
          "example": "sodum"
        },
        "phone": {
          "type": "number",
          "example": 69
        },
        "userName": {
          "type": "string",
          "example": "neverLoose"
        },
        "address": {
          "type": "string",
          "example": "The ice rank 69"
        }
      },
      "required": [
        "firstName",
        "lastName",
        "email",
        "password",
        "phone",
        "userName",
        "address"
      ]
    },
    "postbrand": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string",
          "example": "Ontario sportswear"
        }
      },
      "required": [
        "brand"
      ]
    },
    "alterbrand": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "brand": {
          "type": "string",
          "example": "Ontario sportswear"
        }
      },
      "required": [
        "id",
        "brand"
      ]
    },
    "delete": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "id"
      ]
    },
...
    "postproduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Season tickets"
        },
        "description": {
          "type": "string",
          "example": "Hockey tickets for the NOSHO season"
        },
        "price": {
          "type": "number",
          "example": 499
        },
        "quantity": {
          "type": "number",
          "example": 20
        },
        "imgurl": {
          "type": "string",
          "example": "https://i0.wp.com/puckjunk.com/wp-content/uploads/2024/05/Shoresy-Poster.jpg?ssl=1"
        }
      },
      "required": [
        "name",
        "description",
        "price",
        "quantity",
        "imgurl"
      ]
    },
    "alterproduct": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Season tickets"
        },
        "description": {
          "type": "string",
          "example": "Hockey tickets for the NOSHO season"
        },
        "price": {
          "type": "number",
          "example": 499
        },
        "quantity": {
          "type": "number",
          "example": 20
        },
        "imgurl": {
          "type": "string",
          "example": "https://i0.wp.com/puckjunk.com/wp-content/uploads/2024/05/Shoresy-Poster.jpg?ssl=1"
        }
      },
      "required": [
        "id",
        "name",
        "description",
        "price",
        "quantity",
        "imgurl"
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "token",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": {
    "Admin": [
      "adminJWTtoken"
    ]
  }
}